Extract out new module `AssetId.Gen`.
