+++ lib/coin-selection/test/spec/Cardano/CoinSelection/UTxOIndexSpec.hs
@@ -33,16 +33,16 @@ import Cardano.CoinSelection.UTxOIndex.Internal
 import Cardano.Wallet.Primitive.Types.AssetId
     ( AssetId
     )
+import Cardano.Wallet.Primitive.Types.AssetId.Gen
+    ( genAssetId
+    , shrinkAssetId
+    )
 import Cardano.Wallet.Primitive.Types.TokenBundle
     ( TokenBundle
     )
 import Cardano.Wallet.Primitive.Types.TokenBundle.Gen
     ( genTokenBundleSmallRangePositive
     , shrinkTokenBundleSmallRangePositive
-    )
-import Cardano.Wallet.Primitive.Types.TokenMap.Gen
-    ( genAssetId
-    , shrinkAssetId
     )
 import Control.Monad
     ( void
