+++ lib/primitive/lib/Cardano/Wallet/Primitive/Types/TokenMap/Gen.hs
@@ -1,37 +1,19 @@
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE NamedFieldPuns #-}
-
 module Cardano.Wallet.Primitive.Types.TokenMap.Gen
-    ( genAssetId
-    , genAssetIdLargeRange
-    , genTokenMap
+    ( genTokenMap
     , genTokenMapSmallRange
-    , shrinkAssetId
     , shrinkTokenMap
-    , AssetIdF (..)
     , genTokenMapPartition
     , genTokenMapPartitionNonNull
     ) where
 
 import Prelude
 
-import Cardano.Wallet.Primitive.Types.AssetId
-    ( AssetId (..)
+import Cardano.Wallet.Primitive.Types.AssetId.Gen
+    ( genAssetId
+    , shrinkAssetId
     )
 import Cardano.Wallet.Primitive.Types.TokenMap
     ( TokenMap
-    )
-import Cardano.Wallet.Primitive.Types.TokenName.Gen
-    ( genTokenName
-    , genTokenNameLargeRange
-    , shrinkTokenName
-    , testTokenNames
-    )
-import Cardano.Wallet.Primitive.Types.TokenPolicyId.Gen
-    ( genTokenPolicyId
-    , genTokenPolicyIdLargeRange
-    , shrinkTokenPolicyId
-    , testTokenPolicyIds
     )
 import Cardano.Wallet.Primitive.Types.TokenQuantity
     ( TokenQuantity (..)
@@ -45,61 +27,28 @@ import Control.Monad
     ( replicateM
     )
 import Data.List
-    ( elemIndex
-    , transpose
+    ( transpose
     )
 import Data.List.NonEmpty
     ( NonEmpty
-    )
-import Data.Maybe
-    ( fromMaybe
-    )
-import GHC.Generics
-    ( Generic
     )
 import Safe
     ( fromJustNote
     )
 import Test.QuickCheck
-    ( CoArbitrary (..)
-    , Function (..)
-    , Gen
+    ( Gen
     , choose
-    , functionMap
     , oneof
     , shrinkList
     , sized
-    , variant
     )
 import Test.QuickCheck.Extra
-    ( genSized2With
-    , shrinkInterleaved
+    ( shrinkInterleaved
     )
 
 import qualified Cardano.Wallet.Primitive.Types.TokenMap as TokenMap
 import qualified Data.Foldable as F
 import qualified Data.List.NonEmpty as NE
-
---------------------------------------------------------------------------------
--- Asset identifiers chosen from a range that depends on the size parameter
---------------------------------------------------------------------------------
-
-genAssetId :: Gen AssetId
-genAssetId = genSized2With AssetId genTokenPolicyId genTokenName
-
-shrinkAssetId :: AssetId -> [AssetId]
-shrinkAssetId (AssetId p t) = uncurry AssetId <$> shrinkInterleaved
-    (p, shrinkTokenPolicyId)
-    (t, shrinkTokenName)
-
---------------------------------------------------------------------------------
--- Asset identifiers chosen from a large range (to minimize collisions)
---------------------------------------------------------------------------------
-
-genAssetIdLargeRange :: Gen AssetId
-genAssetIdLargeRange = AssetId
-    <$> genTokenPolicyIdLargeRange
-    <*> genTokenNameLargeRange
 
 --------------------------------------------------------------------------------
 -- Token maps with assets and quantities chosen from ranges that depend on the
@@ -141,22 +90,6 @@ shrinkTokenMap
     shrinkAssetQuantity (a, q) = shrinkInterleaved
         (a, shrinkAssetId)
         (q, shrinkTokenQuantity)
-
---------------------------------------------------------------------------------
--- Filtering functions
---------------------------------------------------------------------------------
-
-newtype AssetIdF = AssetIdF AssetId
-    deriving (Generic, Eq, Show, Read)
-
-instance Function AssetIdF where
-    function = functionMap show read
-
-instance CoArbitrary AssetIdF where
-    coarbitrary (AssetIdF AssetId{tokenName, tokenPolicyId}) genB = do
-        let n = fromMaybe 0 (elemIndex tokenName testTokenNames)
-        let m = fromMaybe 0 (elemIndex tokenPolicyId testTokenPolicyIds)
-        variant (n+m) genB
 
 --------------------------------------------------------------------------------
 -- Partitioning token maps
