+++ lib/wallet/test/unit/Cardano/WalletSpec.hs
@@ -135,6 +135,9 @@ import Cardano.Wallet.Primitive.Types.Address
 import Cardano.Wallet.Primitive.Types.Address.Gen
     ( genAddress
     )
+import Cardano.Wallet.Primitive.Types.AssetId.Gen
+    ( genAssetIdLargeRange
+    )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin (..)
     )
@@ -150,9 +153,6 @@ import Cardano.Wallet.Primitive.Types.RewardAccount
 import Cardano.Wallet.Primitive.Types.TokenBundle
     ( TokenBundle (TokenBundle)
     , getAssets
-    )
-import Cardano.Wallet.Primitive.Types.TokenMap.Gen
-    ( genAssetIdLargeRange
     )
 import Cardano.Wallet.Primitive.Types.TokenQuantity.Gen
     ( genTokenQuantityPositive
